third_party/blink/web_tests/webexposed/internal-properties-should-not-be-exposed-expected.txt

This test checks that -internal-* properties are not exposed to JS.




Testing -internal-marquee-direction
PASS window.getComputedStyle(element).getPropertyValue(property) is ""
PASS element.style[property] is undefined.
Value after setting -internal-marquee-direction to alternate
PASS element.style[property] is undefined.


PASS successfullyParsed is true

TEST COMPLETE
